lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
pyjwt==1.7.1
requests
cryptography